move pipes and file-upload in ui folder
+ remove unused LoadScriptService
