+++ libs/core/src/lib/ui/file-upload/file-drop.directive.ts
