+++ libs/core/src/lib/ui/index.ts
@@ -1,1 +1,4 @@
+export * from './file-upload';
+export * from './pipes';
+export * from './sidebar.service';
 export * from './splash-screen.service';
