+++ libs/core/src/lib/ui/pipes/filter.pipe.ts
