+++ libs/core/src/lib/ui/pipes/floor.pipe.ts
