+++ libs/core/src/lib/ui/pipes/safe.pipe.ts
