+++ libs/core/src/lib/ui/pipes/size.pipe.ts
