+++ libs/core/src/lib/ui/pipes/stf-pipes.module.ts
