+++ libs/core/src/lib/ui/pipes/trim.pipe.ts
